requirements.txt
aiogram==3.5.0
python-dotenv